refactor(dashboard): move the OpenCode page under a coding-agents section
