+++ packages/xinity-ai-dashboard/src/routes/(authenticated)/docs/coding-agents/opencode/+page.server.ts
